+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["common", "native-host", "zkvm-client", "zkvm-host"]
+members = ["crates/*", "native-host", "zkvm-client", "zkvm-host"]
 
 [workspace.package]
 license = "MIT"
@@ -41,7 +41,8 @@ rkyv = "0.7.44"
 hex = "0.4.3"
 
 # local
-zkvm-common = { path = "common" }
+client-utils = { path = "crates/client" }
+host-utils = { path = "crates/host" }
 zkvm-client = { path = "zkvm-client" }
 zkvm-host = { path = "zkvm-host" }
