+++ crates/client/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "zkvm-common"
+name = "client-utils"
 version = "0.1.0"
 license.workspace = true
 edition.workspace = true
