+++ crates/client/src/hasher.rs
