+++ crates/client/src/lib.rs
@@ -2,7 +2,10 @@ mod hasher;
 pub use hasher::BytesHasherBuilder;
 
 mod boot;
-pub use boot::BootInfoWithoutRollupConfig;
+pub use boot::RawBootInfo;
 
 mod executor;
 pub use executor::block_on;
+
+mod oracle;
+use oracle::InMemoryOracle;
