+++ native-host/src/main.rs
@@ -1,9 +1,10 @@
 use clap::Parser;
 use kona_host::HostCli;
-use native_host::run_native_host;
+use kona_host::{init_tracing_subscriber, start_server_and_native_client, HostCli};
 
 #[tokio::main(flavor = "multi_thread")]
 async fn main() {
     let cfg = HostCli::parse();
-    run_native_host(&cfg).await.unwrap();
+    init_tracing_subscriber(cfg.v).unwrap();
+    start_server_and_native_client(cfg.clone()).await.unwrap();
 }
